# api/openapi.yaml
openapi: 3.1.0
info:
  title: MaxPosty API
  version: 1.0.0
  description: |
    Canonical HTTP contract shared by the MaxPosty backend and frontend.
    Browser sessions are carried by the secure `maxstudio_session` cookie.
servers:
  - url: https://api.maxposty.ru/api/v1
    description: Production
  - url: http://localhost:8080/api/v1
    description: Local development
...
paths:
...
  /posts/format-content:
    post:
      tags: [AI]
      operationId: formatPostContent
      summary: Format plain text while preserving meaning
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content, format]
              properties:
                content:
                  type: string
                  minLength: 1
                  maxLength: 4000
                format:
                  $ref: '#/components/schemas/PostFormat'
      responses:
        '200':
          description: Formatted content
          content:
            application/json:
              schema:
                type: object
                required: [content]
                properties:
                  content:
                    type: string
        default:
          $ref: '#/components/responses/Problem'
...
  /workspaces:
    get:
      tags: [Workspaces]
      operationId: listWorkspaces
      summary: List workspaces available to the current user with authoritative capabilities
      security: [{sessionCookie: []}]
      responses:
        '200': {$ref: '#/components/responses/WorkspaceList'}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Workspaces]
      operationId: createWorkspace
      summary: Create a team workspace
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WorkspaceCreate'}
      responses:
        '201': {$ref: '#/components/responses/WorkspaceEnvelope'}
        default: {$ref: '#/components/responses/Problem'}
...
  /workspaces/{workspace_id}/posts/format-content:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}]
    post:
      tags: [AI]
      operationId: formatWorkspacePostContent
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/FormatContentInput'}
      responses:
        '200': {description: Formatted content, content: {application/json: {schema: {type: object}}}}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspace_id}/research/generate:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}]
    post:
      tags: [AI]
      operationId: generateWorkspaceResearch
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/WorkspaceResearchInput'}
      responses:
        '200': {description: Research result, content: {application/json: {schema: {$ref: '#/components/schemas/ResearchResult'}}}}
        default: {$ref: '#/components/responses/Problem'}
...
  /workspaces/{workspace_id}/images/generate:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}]
    post:
      tags: [AI]
      operationId: generateWorkspaceImage
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GenerateImageInput'}
      responses:
        '201': {$ref: '#/components/responses/MediaFile'}
        default: {$ref: '#/components/responses/Problem'}
  /workspaces/{workspace_id}/media:
    parameters: [{$ref: '#/components/parameters/WorkspaceId'}]
    post:
      tags: [Posts]
      operationId: uploadWorkspaceMedia
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: {type: string, format: binary}
      responses:
        '201': {$ref: '#/components/responses/MediaFile'}
        default: {$ref: '#/components/responses/Problem'}
...
  /workspaces/{workspace_id}/posts/{post_id}/generate-image:
    parameters:
      - {$ref: '#/components/parameters/WorkspaceId'}
      - {$ref: '#/components/parameters/PostId'}
    post:
      tags: [AI]
      operationId: generateWorkspacePostImage
      summary: Generate and attach an image to a workspace post
      security: [{sessionCookie: []}]
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/GeneratePostImageInput'}
      responses:
        '200': {$ref: '#/components/responses/Post'}
        default: {$ref: '#/components/responses/Problem'}
...
  /research/generate:
    post:
      tags: [AI]
      operationId: generateResearch
      summary: Research a topic and prepare a post draft
      security:
        - sessionCookie: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ResearchInput'
      responses:
        '200':
          description: Research result
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ResearchResult'
        default:
          $ref: '#/components/responses/Problem'
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: maxstudio_session
  parameters:
...
    WorkspaceId:
      name: workspace_id
      in: path
      required: true
      schema: {type: string, minLength: 1, maxLength: 128}
...
    PostId:
      name: post_id
      in: path
      required: true
      schema: {type: integer, format: int64, minimum: 1}
...
  responses:
    Problem:
      description: Request failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ProblemEnvelope'
    Post:
      description: Updated post
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Post'
...
    WorkspaceEnvelope:
      description: Workspace and authoritative access context
      content: {application/json: {schema: {$ref: '#/components/schemas/WorkspaceEnvelope'}}}
    WorkspaceList:
      description: Available workspaces
      content: {application/json: {schema: {type: array, items: {$ref: '#/components/schemas/WorkspaceEnvelope'}}}}
...
    MediaFile:
      description: Private workspace media metadata
      content: {application/json: {schema: {$ref: '#/components/schemas/MediaFile'}}}
...
  schemas:
    ProblemEnvelope:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/Problem'
    Problem:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
        message:
          type: string
        details: {}
...
    WorkspaceCapability:
      type: string
      enum:
        - workspace.read
        - workspace.update
        - workspace.delete
        - members.read
        - members.manage
        - invitations.read
        - invitations.manage
        - channels.read
        - channels.manage
        - posts.read
        - posts.write
        - posts.delete
        - media.read
        - media.write
        - ai.use
        - comments.read
        - comments.write
        - comments.resolve
        - review.submit
        - review.decide
        - posts.publish
        - audit.read
        - notifications.read
        - notifications.manage
    WorkspaceRole:
      type: string
      enum: [owner, editor, approver, viewer]
    Workspace:
      type: object
      required:
        - id
        - name
        - owner_user_id
        - is_personal
        - approval_required
        - require_distinct_approver
        - created_at
        - updated_at
      properties:
        id: {type: string}
        name: {type: string, minLength: 1, maxLength: 120}
        owner_user_id: {type: string}
        is_personal: {type: boolean}
        approval_required: {type: boolean}
        require_distinct_approver: {type: boolean}
        created_by: {type: string}
        created_at: {type: string, format: date-time}
        updated_at: {type: string, format: date-time}
        archived_at: {type: [string, 'null'], format: date-time}
    WorkspaceAccess:
      type: object
      required: [workspace_id, role, capabilities]
      properties:
        workspace_id: {type: string}
        role: {$ref: '#/components/schemas/WorkspaceRole'}
        capabilities:
          type: array
          uniqueItems: true
          items: {$ref: '#/components/schemas/WorkspaceCapability'}
    WorkspaceEnvelope:
      type: object
      required: [workspace, access]
      properties:
        workspace: {$ref: '#/components/schemas/Workspace'}
        access: {$ref: '#/components/schemas/WorkspaceAccess'}
...
    WorkspaceCreate:
      type: object
      required: [name]
      properties:
        name: {type: string, minLength: 1, maxLength: 120}
...
    FormatContentInput:
      type: object
      required: [content, format]
      properties:
        content: {type: string, minLength: 1, maxLength: 4000}
        format: {$ref: '#/components/schemas/PostFormat'}
    MediaFile:
      type: object
      required: [type, url, filename, mime_type, size, width, height]
      properties:
        type: {type: string, enum: [image, video]}
        url: {type: string}
        filename: {type: string}
        mime_type: {type: string}
        size: {type: integer, format: int64, minimum: 0}
        width: {type: integer, minimum: 0}
        height: {type: integer, minimum: 0}
        duration_ms: {type: integer, format: int64, minimum: 0}
...
    PostReviewStatus:
      type: string
      enum: [draft, in_review, changes_requested, approved]
...
    PostFormat:
      type: string
      enum: [markdown, html]
    PostStatus:
      type: string
      enum: [draft, scheduled, publishing, published, failed]
    PostLinkButton:
      type: object
      required: [text, url]
      properties:
        text:
          type: string
          minLength: 1
          maxLength: 64
        url:
          type: string
          format: uri
    PostAttachment:
      type: object
      required:
        - id
        - type
        - position
        - url
        - processing_status
        - size_bytes
        - mime_type
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        type:
          type: string
          enum: [image, video]
        position:
          type: integer
          minimum: 0
          maximum: 11
        url:
          type: string
        processing_status:
          type: string
          enum: [uploading, processing, ready, failed]
        size_bytes:
          type: integer
          format: int64
          minimum: 0
        mime_type:
          type: string
        width:
          type: [integer, 'null']
          minimum: 1
        height:
          type: [integer, 'null']
          minimum: 1
        duration_ms:
          type: [integer, 'null']
          format: int64
          minimum: 0
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Post:
      type: object
      required:
        - id
        - title
        - content
        - format
        - status
        - attachments
        - link_buttons
        - notify
        - disable_link_preview
        - max_is_pinned
        - created_at
        - updated_at
      properties:
        id:
          type: integer
          format: int64
        workspace_id:
          type: string
        title:
          type: string
        content:
          type: string
        format:
          $ref: '#/components/schemas/PostFormat'
        status:
          $ref: '#/components/schemas/PostStatus'
        channel_id:
          type: [integer, 'null']
          format: int64
        image_url:
          type: [string, 'null']
        image_prompt:
          type: [string, 'null']
        attachments:
          type: array
          maxItems: 12
          items:
            $ref: '#/components/schemas/PostAttachment'
        link_buttons:
          type: array
          items:
            $ref: '#/components/schemas/PostLinkButton'
        notify:
          type: boolean
        disable_link_preview:
          type: boolean
        scheduled_at:
          type: [string, 'null']
          format: date-time
        max_message_id:
          type: [string, 'null']
        max_message_url:
          type: [string, 'null']
          format: uri
        max_views:
          type: [integer, 'null']
          format: int64
          minimum: 0
        max_stats_synced_at:
          type: [string, 'null']
          format: date-time
        max_is_pinned:
          type: boolean
        last_error:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        published_at:
          type: [string, 'null']
          format: date-time
        review_status:
          $ref: '#/components/schemas/PostReviewStatus'
        current_revision_id:
          type: [integer, 'null']
          format: int64
    GenerateImageInput:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 2000
        size:
          type: string
          enum: [1024x1024, 1536x1024, 1024x1536]
          default: 1024x1024
        quality:
          type: string
          enum: [low, medium, high]
          default: medium
    GeneratePostImageInput:
      type: object
      properties:
        prompt:
          type: string
          maxLength: 2000
          description: Optional override; a blank or omitted value uses the post image prompt.
        size:
          type: string
          enum: [1024x1024, 1536x1024, 1024x1536]
          default: 1024x1024
        quality:
          type: string
          enum: [low, medium, high]
          default: medium
...
    ResearchInput:
      type: object
      required: [topic, tone, format, include_sources]
      properties:
        topic:
          type: string
          minLength: 1
          maxLength: 500
        angle:
          type: string
        audience:
          type: string
        tone:
          type: string
          maxLength: 100
        cta:
          type: string
          maxLength: 500
        forbidden_words:
          type: array
          maxItems: 50
          items: {type: string, maxLength: 100}
        example_posts:
          type: array
          maxItems: 10
          items: {type: string, maxLength: 4000}
        visual_style:
          type: string
          maxLength: 1000
        format:
          $ref: '#/components/schemas/PostFormat'
        include_sources:
          type: boolean
    WorkspaceResearchInput:
      allOf:
        - $ref: '#/components/schemas/ResearchInput'
        - type: object
          properties:
            channel_template_id: {type: integer, format: int64, minimum: 1}
            channel_id: {type: integer, format: int64, minimum: 1}
...
    ResearchSource:
      type: object
      required: [title, url]
      properties:
        title:
          type: string
        url:
          type: string
          format: uri
    ResearchDraft:
      type: object
      required: [title, content, format, image_prompt]
      properties:
        title:
          type: string
        content:
          type: string
        format:
          $ref: '#/components/schemas/PostFormat'
        image_prompt:
          type: string
    ResearchResult:
      type: object
      required: [topic, report, sources, draft]
      properties:
        topic:
          type: string
        report:
          type: string
        draft:
          $ref: '#/components/schemas/ResearchDraft'
        sources:
          type: array
          items:
            $ref: '#/components/schemas/ResearchSource'
...